refactor: move LLM utilities to shared module
